Changed folder sturcture
Keep all the screens in `view`
