--- lib/constants/lists.dart
@@ -1,2 +1,0 @@
-const List popUpList = ["Manage Categories","Sort By", "Upgrade to Pro"];
-const List categoryList = ["Hide", "Delete", "Edit"];
