--- lib/models/category.dart
@@ -1,8 +1,0 @@
-
-class Category {
-  int id;
-  Category({
-    required this.id,
-  });
-  
-}
