+++ lib/shared/theme/extentions.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_todo/theme/dark.dart';
-import 'package:flutter_todo/theme/light.dart';
-import 'package:flutter_todo/theme/text.dart';
+import 'package:flutter_todo/shared/theme/dark.dart';
+import 'package:flutter_todo/shared/theme/light.dart';
+import 'package:flutter_todo/shared/theme/text.dart';
 
 extension TextThemeExtention on ColorScheme {
   CustomTextFields get customTextStyle => CustomTextFields(
