+++ lib/view/sign_in/components/body.dart
@@ -4,30 +4,24 @@ import 'package:flutter/material.dart';
 import 'package:flutter_keyboard_visibility/flutter_keyboard_visibility.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:flutter_todo/animations/animations.dart';
-import 'package:flutter_todo/authentication/sign_in/sign_in.dart';
 import 'package:flutter_todo/shared/components/customProceedButton.dart';
 import 'package:flutter_todo/shared/components/customTextField.dart';
-import 'package:flutter_todo/shared/validitors.dart';
-import 'package:flutter_todo/view_model/sign_up.dart';
+import 'package:flutter_todo/view/sign_up/sign_up.dart';
 
-import 'package:get/get.dart';
-
-class SignUpScreenBody extends StatefulWidget {
-  const SignUpScreenBody({Key? key}) : super(key: key);
+class SignInScreenBody extends StatefulWidget {
+  const SignInScreenBody({Key? key}) : super(key: key);
 
   @override
-  State<SignUpScreenBody> createState() => _SignUpScreenBodyState();
+  State<SignInScreenBody> createState() => _SignInScreenBodyState();
 }
 
-class _SignUpScreenBodyState extends State<SignUpScreenBody> {
+class _SignInScreenBodyState extends State<SignInScreenBody> {
   bool iskeyboardVisivle = false;
   late final StreamSubscription keyboardStream;
   @override
   void initState() {
     super.initState();
     final keyboardVisibilityController = KeyboardVisibilityController();
-
-    // iskeyboardVisivle = true;
 
     keyboardStream =
         keyboardVisibilityController.onChange.listen((bool visible) {
@@ -41,8 +35,6 @@ class _SignUpScreenBodyState extends State<SignUpScreenBody> {
     keyboardStream.cancel();
     super.dispose();
   }
-
-  SignUpController controller = Get.find();
 
   @override
   Widget build(BuildContext context) {
@@ -60,77 +52,58 @@ class _SignUpScreenBodyState extends State<SignUpScreenBody> {
                   ? 50
                   : MediaQuery.of(context).size.height * 0.3,
               child: AnimatedContainer(
-                curve: Curves.linear,
+                curve: Curves.decelerate,
                 duration: const Duration(milliseconds: 200),
                 height: iskeyboardVisivle
                     ? 0
                     : MediaQuery.of(context).size.height * 0.3,
                 child: AnimatedOpacity(
-                  child: SvgPicture.asset('assets/svg/sign_up.svg'),
+                  child: SvgPicture.asset('assets/svg/sign_in.svg'),
                   duration: const Duration(milliseconds: 200),
-                  curve: Curves.linear,
+                  curve: Curves.decelerate,
                   opacity: iskeyboardVisivle ? 0 : 1,
                 ),
               ),
             ),
             const SizedBox(height: 20),
-            Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                const Text(
-                  'Enter Your Name',
-                  style: TextStyle(
-                    color: Colors.black,
-                    fontFamily: 'Metro',
-                    fontSize: 30,
-                    fontWeight: FontWeight.w700,
-                  ),
-                ),
-                const SizedBox(height: 20),
-                Form(
-                  key: controller.nameScreenFormKey,
-                  child: Column(
-                    children: [
-                      CustomTextField(
-                        controller: controller.firstNameController,
-                        hint: 'First Name',
-                        textFieldType: CustomTextFieldType.regular,
-                        node: controller.firstNameFocusNode,
-                        validator: (value) => lengthValidator(value, 3),
-                        nextNode: controller.lastNameFocusNode,
-                      ),
-                      const SizedBox(height: 5),
-                      CustomTextField(
-                        controller: controller.lastNameController,
-                        hint: 'Last Name',
-                        textFieldType: CustomTextFieldType.regular,
-                        node: controller.lastNameFocusNode,
-                        onFieldSubmitted: (p0) => controller.nextTab(context),
-                      ),
-                    ],
-                  ),
-                ),
-              ],
+            const Text(
+              'Sign In',
+              style: TextStyle(
+                color: Colors.black,
+                fontFamily: 'Metro',
+                fontSize: 30,
+                fontWeight: FontWeight.w700,
+                letterSpacing: 1.2,
+              ),
             ),
-
             const SizedBox(height: 20),
-            CustomProceedButton(
-              title: 'Next',
-              heightFactor: 1,
-              onPressed: () => controller.nextTab(context),
+            const CustomTextField(
+              hint: 'Email',
+              textFieldType: CustomTextFieldType.email,
+            ),
+            const SizedBox(height: 5),
+            const CustomTextField(
+              hint: 'Password',
+              textFieldType: CustomTextFieldType.password,
+            ),
+            const SizedBox(height: 20),
+
+            const CustomProceedButton(
+              title: 'Login',
+              heightFactor: .9,
             ),
             const SizedBox(height: 10),
             Center(
               child: GestureDetector(
                 onTap: () => Navigator.of(context).pushReplacement(
                   PageTransition(
-                    child: const SignInScreen(),
-                    type: PageTransitionType.fromLeft,
+                    child: SignUpScreen(),
+                    type: PageTransitionType.fromRight,
                   ),
                 ),
                 child: Text.rich(
                   TextSpan(
-                    text: 'Already have an account? ',
+                    text: 'Don\'t have an account? ',
                     style: TextStyle(
                       color: Colors.black.withOpacity(.6),
                       fontFamily: 'Metro',
@@ -139,7 +112,7 @@ class _SignUpScreenBodyState extends State<SignUpScreenBody> {
                     ),
                     children: const [
                       TextSpan(
-                        text: 'Sign In',
+                        text: 'Sign Up',
                         style: TextStyle(
                           color: Colors.black,
                           fontFamily: 'Metro',
@@ -152,7 +125,18 @@ class _SignUpScreenBodyState extends State<SignUpScreenBody> {
                 ),
               ),
             ),
-
+            const SizedBox(height: 5),
+            const Center(
+              child: Text(
+                'Forgot Password?',
+                style: TextStyle(
+                  color: Colors.blue,
+                  fontFamily: 'Metro',
+                  fontSize: 15,
+                  fontWeight: FontWeight.w700,
+                ),
+              ),
+            ),
             // if (!iskeyboardVisivle)
             // const SizedBox(height: 100),
           ],
