+++ lib/view/sign_in/sign_in.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_todo/authentication/sign_in/components/body.dart';
+import 'package:flutter_todo/view/sign_in/components/body.dart';
 
 class SignInScreen extends StatelessWidget {
   const SignInScreen({Key? key}) : super(key: key);
