+++ lib/view/sign_up/components/complete_setup.dart
