+++ lib/view/welcome/components/body.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_todo/animations/animations.dart';
-import 'package:flutter_todo/authentication/welcome/components/auth_options.dart';
-import 'package:flutter_todo/authentication/welcome/components/intro.dart';
-import 'package:flutter_todo/authentication/welcome/components/logo.dart';
-import 'package:flutter_todo/authentication/welcome/components/sign_up.dart';
+import 'package:flutter_todo/view/welcome/components/auth_options.dart';
+import 'package:flutter_todo/view/welcome/components/intro.dart';
+import 'package:flutter_todo/view/welcome/components/logo.dart';
+import 'package:flutter_todo/view/welcome/components/sign_up.dart';
 
 class WelcomeScreenBody extends StatelessWidget {
   const WelcomeScreenBody({Key? key}) : super(key: key);
