+++ lib/view/welcome/components/intro.dart
