+++ lib/view/welcome/components/logo.dart
