+++ lib/view/welcome/welcome.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_todo/authentication/welcome/components/body.dart';
+import 'package:flutter_todo/view/welcome/components/body.dart';
 
 class WelcomeScreen extends StatelessWidget {
   const WelcomeScreen({Key? key}) : super(key: key);
